Late submition, it did not went through last night.
I uploaded last night but this morning realized that it did not went through. The resampling is completed, but i am still working on debugging it. I am also working on the ensemble file. 

Apologize for the late submittion but i am use to study in a slower pace to really digest the new information, and i am struggling with that since the amount of new information in every class is overwhelming. I will keep working on these files either way.